requirements.txt
Django==1.11.3
django-makemessages-xgettext==0.1.1
django-compressor==2.1.1
django-libsass==0.7
django-modeltranslation==0.12.1
django-imagekit==4.0.1
libsass==0.13.2
psycopg2==2.7.1
pytz==2017.2
Pillow==3.3.2

